已完成：登录、注册，加密算法

+++ api-common/src/main/java/top/twip/common/entity/user/WebsiteUserInfo.java
@@ -21,8 +21,8 @@ public class WebsiteUserInfo {
     private Integer isadmin;
 
     @TableField(fill = FieldFill.INSERT)
-    private Integer ctime;
+    private Long ctime;
 
     @TableField(fill = FieldFill.INSERT_UPDATE)
-    private Integer mtime;
+    private Long mtime;
 }
